backend/pyproject.toml
[project]
name = "steeltrack-backend"
version = "1.1.0"
description = "Modernes Lagerverwaltungssystem für Stahlprodukte mit FastAPI, SQLite und IRIS-Integration."
readme = "../README.md"
requires-python = ">=3.10"
authors = [
    {name = "Nodar Abramishvili", email = "office@softmaster.at"}
]

# Nur die Pakete, die wir direkt im Code benutzen (Imports)
dependencies = [
    "fastapi==0.110.0",
    "uvicorn==0.27.1",
    "sqlalchemy==2.0.27",
    "pydantic==2.6.1",
    "python-multipart==0.0.9",
    "python-dotenv==1.2.2",
    "tinydb==4.8.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
# Tools für die Entwicklung, die nicht auf den Server müssen
dev = [
    "pdoc==16.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
